Add files via upload
The race data used in this project was collected using a modified version of the rpscrape tool which is originally developed by joenano.
I adapted parts of the scraper to suit my own requirements (e.g. adjusting file paths, adding retry behaviour and data-saving logic) but the core structure remains based on the original tool. The rpscrape source code is not included in this repository but it is executed locally to produce the CSV outputs.